+++ examples/flow-react-example/src/OnClick2.gen.js
@@ -1,0 +1,13 @@
+/** 
+ * @flow strict
+ * @generated from OnClick2.res
+ * @nolint
+ */
+/* eslint-disable */
+
+// $FlowExpectedError: Reason checked type sufficiently
+import * as OnClick2BS from './OnClick2.bs';
+
+export type Props = {| +onClick: (MouseEvent) => void |};
+
+export const make: React$ComponentType<{| +onClick: (MouseEvent) => void |}> = OnClick2BS.make;